hoan thien 10/10

+++ app/Http/Controllers/Admin/UserController.php
@@ -8,6 +8,7 @@ use App\Http\Requests\ManageUser;
 use Illuminate\Support\Facades\DB;
 use App\Role;
 use App\User;
+use App\Order;
 
 class UserController extends Controller
 {
@@ -18,7 +19,7 @@ class UserController extends Controller
      */
     public function index()
     {
-        $users = User::paginate(5);
+        $users = User::orderBy('id', 'desc')->paginate(5);
         $roles = Role::all();
         return view('admin.users.index', compact('users', 'roles'));
     }
@@ -39,9 +40,42 @@ class UserController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(ManageUser $request)
+    public function store(Request $request)
     {
-        $data = $request->all();
+        $data = $request->validate([
+            'username' => 'required|max:255',
+            'email' => 'required|email|unique:users,email',
+            'password' => 'required|min:6',
+            'confirm_password' => 'required|same:password',
+            'yourname' => 'required',
+            'phone' => 'required|numeric',
+            'address' => 'required',
+        ],
+        [
+            'username.required' => 'Tên đăng nhập không được để trống',
+            'username.max' => 'Tên đăng nhập không được quá 255 ký tự',
+            'email.required' => 'Email đăng nhập không được để trống',
+            'email.email' => 'Email phải đúng định dạng',
+            'email.unique' => 'Email đăng nhập không được trùng',
+            'password.required' => 'Mật khẩu không được để trống',
+            'password.min' => 'Mật khẩu không được nhỏ hơn 6 ký tự',
+            'confirm_password.same' => 'Mật khẩu không trùng khớp',
+            'yourname.required' => 'Họ và tên không được để trống',
+            'phone.required' => 'Số điện thoại không được để trống',
+            'phone.numeric' => 'Số điện thoại phải là ký tự số',
+            'address.required' => 'Địa chỉ không được để trống',        
+        ]
+        );
+
+        $data = [
+            'username' => $request->username,
+            'email' => $request->email,
+            'password' => bcrypt($request->password),
+            'yourname' => $request->yourname,
+            'phone' => $request->phone, 
+            'address' => $request->address,
+            'role_id' => $request->role_id
+        ];
         User::create($data);
         $request->session()->flash('status', 'Thêm người dùng thành công!');
         return redirect()->route('users.index');
@@ -83,15 +117,12 @@ class UserController extends Controller
         $data = $request->all();
         $data = $request->validate([
             'username' => 'required',
-            'password' => 'required|min:6',
             'yourname' => 'required',
             'phone' => 'required',
             'address' => 'required',
         ],
         [
             'username.required' => 'Tên đăng nhập không được để trống',
-            'password.min' => 'Mật khẩu tối thiểu 6 kí tự',
-            'password.required' => 'Mật khẩu không được để trống',
             'yourname.required' => 'Họ và tên không được để trống',
             'phone.required' => 'Số điện thoại không được để trống',
             'address.required' => 'Địa chỉ không được để trống',        
@@ -112,17 +143,22 @@ class UserController extends Controller
     {
         // $user->delete();
         // return redirect()->route('users.index');
-        $user = User::findOrFail($request->user_id);
-        $user->delete();
-        $request->session()->flash('status', 'Xóa thành công!');
-        return redirect()->route('users.index');
+        if($order_id = Order::where('user_id', $request->user_id)->first()) {
+            $request->session()->flash('error', 'Không được xóa người dùng đang đặt hàng');
+            return redirect()->route('users.index');
+        } else {
+            $user = User::findOrFail($request->user_id);
+            $user->delete();
+            $request->session()->flash('status', 'Xóa thành công!');
+            return redirect()->route('users.index');
+        }
     }
 
     public function search(Request $request)
     {
-        if($request->ajax()){
-            $output="";
-            $users = User::where('username','LIKE','%'.$request->search."%")->get();
+       if ($request->ajax()) {
+            $output = "";
+            $users = DB::table('users')->where('username', 'LIKE', '%'.$request->search."%")->get();
             if($users) {
                 foreach ($users as $key => $user) {
                 $output.='<tr>'.
